implemented Peggleboz; also factored more out to Field
